requirements.txt
asgiref==3.5.2
astroid==2.2.5
autopep8==1.4.4
backports.zoneinfo==0.2.1
certifi==2019.6.16
chardet==3.0.4
defusedxml==0.6.0
Django==2.2.1
django-allauth==0.39.1
django-cors-headers==3.0.2
django-rest-auth==0.9.5
djangorestframework==3.9.4
gunicorn==19.9.0
idna==2.8
isort==4.3.21
lazy-object-proxy==1.4.1
mccabe==0.6.1
oauthlib==3.0.1
pep8==1.7.1
pycodestyle==2.5.0
pylint==2.3.1
python3-openid==3.1.0
pytz==2019.1
requests==2.22.0
requests-oauthlib==1.2.0
six==1.12.0
sqlparse==0.3.0
typed-ast==1.4.0
urllib3==1.25.3
wrapt==1.11.2
